# GFBackend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Gator Forum Backend Server",
        "title": "Gator Forum Backend API",
        "termsOfService": "https://github.com/fongziyjun16/SE/tree/backend",
        "contact": {},
        "version": "1.0"
    },
    "host": "http://167.71.166.120:10010",
    "basePath": "/gf/api",
    "paths": {
...
        "/user/GetUserInfoByUsername": {
            "get": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Manage"
                ],
                "summary": "Get User's Info",
                "responses": {
                    "200": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Search Successfully",
                        "schema": {
                            "$ref": "#/definitions/entity.UserInfo"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    }
                }
            }
        },
        "/user/admin/delete": {
            "post": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need strings username in post request, need token in cookie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Manage"
                ],
                "summary": "Admin delete Users, cannot self delete",
                "parameters": [
                    {
                        "description": "username in post request body",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Update Password Successfully",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    }
                }
            }
        },
        "/user/admin/register": {
            "post": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "only need strings username \u0026 password \u0026 ForAdmin, need token in cookie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Manage"
                ],
                "summary": "Register a new Admin User",
                "parameters": [
                    {
                        "description": "Admin User Register only needs Username, Password(encoded by md5) \u0026 ForAdmin with true.",
                        "name": "UserInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.UserInfo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. User Register Successfully",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters or User Has Existed",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    }
                }
            }
        },
        "/user/follow": {
            "post": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie, need username who is followed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Manage"
                ],
                "summary": "User Follow other users",
                "parameters": [
                    {
                        "description": "username in post request body",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Follow Successfully",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters or User not exist or User has followed.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    }
                }
            }
        },
        "/user/followees": {
            "post": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Manage"
                ],
                "summary": "Get User's followees",
                "parameters": [
                    {
                        "description": "username in post request body",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Search Successfully",
                        "schema": {
                            "$ref": "#/definitions/entity.UserFollows"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    }
                }
            }
        },
        "/user/followers": {
            "post": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Manage"
                ],
                "summary": "Get User's followers",
                "responses": {
                    "201": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Search Successfully",
                        "schema": {
                            "$ref": "#/definitions/entity.UserFollows"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    }
                }
            }
        },
        "/user/getusersinfo": {
            "get": {
                "security": [
                    {
                        "ApiAuthToken": []
                    }
                ],
                "description": "need token in cookie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Manage"
                ],
                "summary": "Get Users' Info By username, pageNo, pageSize, \"/getusersinfo?username=\u0026pageNo=\u0026pageSize=\"",
                "responses": {
                    "200": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. Search Successfully",
                        "schema": {
                            "$ref": "#/definitions/entity.UsersInfo"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "only need strings username \u0026 password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Manage"
                ],
                "summary": "Admin / Regular User login",
                "parameters": [
                    {
                        "description": "only needs username and password",
                        "name": "UserInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.UserInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "\u003cb\u003eSuccess\u003c/b\u003e. User Login Successfully",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "400": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Bad Parameters or Username / Password incorrect",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    },
                    "500": {
                        "description": "\u003cb\u003eFailure\u003c/b\u003e. Server Internal Error.",
                        "schema": {
                            "$ref": "#/definitions/entity.ResponseMsg"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.ResponseMsg": {
            "type": "object",
            "properties": {
                "Nickname": {
                    "type": "string",
                    "example": "James Bond"
                },
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "example": "process successfully"
                },
                "new_community_id": {
                    "type": "integer"
                }
            }
        },
...
        "entity.UserFollows": {
            "type": "object",
            "properties": {
                "Nickname": {
                    "type": "string",
                    "example": "James Bond"
                },
                "Users": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "\"spriderman\"",
                        "\"batman\"",
                        "\"ironman\""
                    ]
                },
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string",
                    "example": "process successfully"
                },
                "new_community_id": {
                    "type": "integer"
                }
            }
        },
        "entity.UserInfo": {
            "type": "object",
            "properties": {
                "ForAdmin": {
                    "type": "boolean"
                },
                "NewPassword": {
                    "type": "string",
                    "example": "3ecb441b741bcd433288f5557e4b9118"
                },
                "Password": {
                    "type": "string",
                    "example": "f9ae5f68ae1e7f7f3fc06053e9b9b539"
                },
                "Username": {
                    "type": "string",
                    "example": "jamesbond21"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiAuthToken": {
            "type": "apiKey",
            "name": "token",
            "in": "cookies"
        }
    }
}